Edit Long isEqual NSIntegaer

+++ Demo/相册Demo/Example/Example0ViewController.m
@@ -86,7 +86,7 @@
 }
 
 #pragma mark - <UITableViewDataSource>
-- (long) tableView:(UITableView *)tableView numberOfRowsInSection:(long)section{
+- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
     return self.assets.count;
 }
