+++ Demo/相册Demo/Example/Example3ViewController.m
@@ -89,7 +89,7 @@
 }
 
 #pragma mark - <UITableViewDataSource>
-- (long) tableView:(UITableView *)tableView numberOfRowsInSection:(long)section{
+- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
     return self.assets.count;
 }
