+++ Demo/相册Demo/Example/Example5ViewController.m
@@ -98,7 +98,7 @@
 
 
 #pragma mark - <UITableViewDataSource>
-- (long) tableView:(UITableView *)tableView numberOfRowsInSection:(long)section{
+- (NSInteger) tableView:(UITableView *)tableView numberOfRowsInSection:(NSInteger)section{
     return self.assets.count;
 }
