+++ Demo/相册Demo/Example/Example7ViewController.m
@@ -64,10 +64,10 @@
     NSUInteger assetCount = self.assets.count + 1;
     
     CGFloat width = self.view.frame.size.width / column;
-    for (long i = 0; i < assetCount; i++) {
+    for (NSInteger i = 0; i < assetCount; i++) {
         
-        long row = i / column;
-        long col = i % column;
+        NSInteger row = i / column;
+        NSInteger col = i % column;
         
         UIButton *btn = [UIButton buttonWithType:UIButtonTypeCustom];
         btn.imageView.contentMode = UIViewContentModeScaleAspectFill;
@@ -126,11 +126,11 @@
 
 
 #pragma mark - <ZLPhotoPickerBrowserViewControllerDataSource>
-- (long)numberOfSectionInPhotosInPickerBrowser:(ZLPhotoPickerBrowserViewController *)pickerBrowser{
+- (NSInteger)numberOfSectionInPhotosInPickerBrowser:(ZLPhotoPickerBrowserViewController *)pickerBrowser{
     return 1;
 }
 
-- (long)photoBrowser:(ZLPhotoPickerBrowserViewController *)photoBrowser numberOfItemsInSection:(NSUInteger)section{
+- (NSInteger)photoBrowser:(ZLPhotoPickerBrowserViewController *)photoBrowser numberOfItemsInSection:(NSUInteger)section{
     return self.assets.count;
 }
