+++ ZLPhotoLib/Classes/ZLPhotoBrowser/ViewControllers/ZLPhotoPickerBrowserViewController.m
@@ -30,7 +30,7 @@ static NSString *_cellIdentifier = @"collectionViewCell";
 // 单击时执行销毁的block
 @property (nonatomic , copy) ZLPickerBrowserViewControllerTapDisMissBlock disMissBlock;
 // 当前提供的分页数
-@property (nonatomic , assign) long currentPage;
+@property (nonatomic , assign) NSInteger currentPage;
 @end
 
 
@@ -208,7 +208,7 @@ static NSString *_cellIdentifier = @"collectionViewCell";
     }
     
     __weak typeof(self)weakSelf = self;
-    self.disMissBlock = ^(long page){
+    self.disMissBlock = ^(NSInteger page){
         mainView.hidden = NO;
         mainView.alpha = 1.0;
         CGRect originalFrame = CGRectZero;
@@ -379,9 +379,9 @@ static NSString *_cellIdentifier = @"collectionViewCell";
 #pragma mark getPhotos
 - (NSArray *) getPhotos{
     NSMutableArray *photos = [NSMutableArray array];
-    long section = self.currentIndexPath.section;
-    long rows = [self.dataSource photoBrowser:self numberOfItemsInSection:section];
-    for (long i = 0; i < rows; i++) {
+    NSInteger section = self.currentIndexPath.section;
+    NSInteger rows = [self.dataSource photoBrowser:self numberOfItemsInSection:section];
+    for (NSInteger i = 0; i < rows; i++) {
         [photos addObject:[self.dataSource photoBrowser:self photoAtIndexPath:[NSIndexPath indexPathForItem:i inSection:section]]];
     }
     return photos;
@@ -444,11 +444,11 @@ static NSString *_cellIdentifier = @"collectionViewCell";
 }
 
 #pragma mark - <UICollectionViewDataSource>
-- (long) numberOfSectionsInCollectionView:(UICollectionView *)collectionView{
+- (NSInteger) numberOfSectionsInCollectionView:(UICollectionView *)collectionView{
     return 1;
 }
 
-- (long) collectionView:(UICollectionView *)collectionView numberOfItemsInSection:(long)section{
+- (NSInteger) collectionView:(UICollectionView *)collectionView numberOfItemsInSection:(NSInteger)section{
     if ([self isDataSourceElsePhotos]) {
         return [self.dataSource photoBrowser:self numberOfItemsInSection:self.currentIndexPath.section];
     }
@@ -517,7 +517,7 @@ static NSString *_cellIdentifier = @"collectionViewCell";
 #pragma mark - <UIScrollViewDelegate>
 - (void)scrollViewDidScroll:(UIScrollView *)scrollView{
     CGRect tempF = self.collectionView.frame;
-    long currentPage = (long)((scrollView.contentOffset.x / scrollView.frame.size.width) + 0.5);
+    NSInteger currentPage = (NSInteger)((scrollView.contentOffset.x / scrollView.frame.size.width) + 0.5);
     if (tempF.size.width < [UIScreen mainScreen].bounds.size.width){
         tempF.size.width = [UIScreen mainScreen].bounds.size.width;
     }
@@ -539,13 +539,13 @@ static NSString *_cellIdentifier = @"collectionViewCell";
     self.collectionView.frame = tempF;
 }
 
--(void)setPageLabelPage:(long)page{
+-(void)setPageLabelPage:(NSInteger)page{
     self.pageLabel.text = [NSString stringWithFormat:@"%ld / %ld",page + 1, self.photos.count];
 }
 
 - (void)scrollViewDidEndDecelerating:(UIScrollView *)scrollView{
     
-    long currentPage = (long)(scrollView.contentOffset.x / (scrollView.frame.size.width));
+    NSInteger currentPage = (NSInteger)(scrollView.contentOffset.x / (scrollView.frame.size.width));
     
     if (currentPage == self.photos.count - 2) {
         currentPage = roundf((scrollView.contentOffset.x) / (scrollView.frame.size.width));
@@ -593,9 +593,9 @@ static NSString *_cellIdentifier = @"collectionViewCell";
 }
 
 #pragma mark - <UIAlertViewDelegate>
-- (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(long)buttonIndex{
+- (void)alertView:(UIAlertView *)alertView clickedButtonAtIndex:(NSInteger)buttonIndex{
     if (buttonIndex == 1) {
-        long page = self.currentPage;
+        NSInteger page = self.currentPage;
         if ([self.delegate respondsToSelector:@selector(photoBrowser:removePhotoAtIndexPath:)]) {
             [self.delegate photoBrowser:self removePhotoAtIndexPath:[NSIndexPath indexPathForItem:page inSection:self.currentIndexPath.section]];
         }
