+++ ZLPhotoLib/Classes/ZLPhotoPicker/Views/ZLPhotoPickerFooterCollectionReusableView.m
@@ -28,11 +28,11 @@
     return _footerLabel;
 }
 
-- (void)setCount:(long)count{
+- (void)setCount:(NSInteger)count{
     _count = count;
     
     if (count > 0) {
-        self.footerLabel.text = [NSString stringWithFormat:@"有 %ld 张图片", (long)count];
+        self.footerLabel.text = [NSString stringWithFormat:@"有 %ld 张图片", (NSInteger)count];
     }
 }
